My Tunes Project- (Completing CRUD Application)
Its more or less working. On The last two genres "Hip-Hop" and
"Neo-Soul" I've put in tracks for the two albums but I can't seem to get
them to show up on the HTML page.  What's weird is that I was able to
apply the tracks in MySQL. Perhaps something is Cashed...